microservices: common: ShellCommandExecutor refactoring
